+++ src/main/php/io/streams/compress/Bzip2.class.php
@@ -22,7 +22,7 @@ class Bzip2 implements Algorithm {
   }
 
   /** Opens an output stream for writing */
-  public function create(OutputStream $out, int $method): OutputStream {
+  public function create(OutputStream $out, int $method= Compression::DEFAULT): OutputStream {
     static $levels= [Compression::FASTEST => 1, Compression::DEFAULT => 4, Compression::STRONGEST => 9];
 
     return new Bzip2OutputStream($out, $levels[$method]);